calculate number of elements

+++ 0x13-more_singly_linked_lists/1-listint_len.c
@@ -1,13 +1,14 @@
 #include "lists.h"
 
-/** listint_len - calculate number of elemets in a linked list
+/**
+ * listint_len -  calculate number of elements in a list
  * @h: list
  *
- * Retrun: number of elements
+ * Return: number of elements
  */
 size_t listint_len(const listint_t *h)
 {
-	unsigned int count = 0;
+	size_t count = 0;
 
 	while (h)
 	{
